provider-validator/requirements.txt
faker
pandas
phonenumbers
python-Levenshtein
requests
aiohttp
beautifulsoup4
pytesseract
pdf2image
python-dotenv
fastapi
uvicorn
email-validator
jinja2
reportlab

fastapi==0.95.2
uvicorn[standard]==0.22.0
pandas
aiohttp
beautifulsoup4
python-dotenv
sqlalchemy==2.0.0
alembic
psycopg2-binary
celery==5.3.4
redis
prometheus-client
opentelemetry-api
opentelemetry-sdk
python-jose
passlib[bcrypt]
phonenumbers
pillow
pytesseract
pdf2image
python-multipart
